Move defineOptions out of configuration.py

+++ olof/configuration.py
@@ -23,44 +23,27 @@ class Configuration(object):
     """
     Store all the configuration options and retrieve the value of a certain option with the ConfigurationParser.
     """
-    def __init__(self, server, configfile):
+    def __init__(self, server, name, defineOptions=None):
         """
         Initialisation. Construct an empty list of options, and fill it with all the Options.
 
         @param  configfile   URL of the configfile to write to.
         """
         self.server = server
+        self.name = name
         self.options = []
-        self.configfile = configfile
-        self._defineOptions()
+        self.configfile = 'olof/config/%s.conf' % name
+        if defineOptions == None:
+            self.options.extend(self._defineOptions())
+        else:
+            self.options.extend(defineOptions())
         self.configparser = _ConfigurationParser(self)
 
     def _defineOptions(self):
         """
         Create all options and add them to the list.
         """
-        ssl_server_crt = _Option(name = 'ssl_server_crt',
-            description = 'Path to the SSL server certificate.',
-            values = {},
-            default = 'keys/server.crt')
-
-        ssl_server_key = _Option(name = 'ssl_server_key',
-            description = 'Path to the SSL server key.',
-            values = {},
-            default = 'keys/server.key')
-
-        ssl_server_ca = _Option(name = 'ssl_server_ca',
-            description = 'Path to the SSL server CA.',
-            values = {},
-            default = 'keys/ca.pem')
-
-        tcp_listening_port = _Option(name = 'tcp_listening_port',
-            description = 'The TCP port to listen on for incoming connection from the scanners.',
-            type = 'int("%s")',
-            values = {2583: 'Listen on port 2583 by default.'},
-            default = 2583)
-
-        self.options.extend([ssl_server_ca, ssl_server_crt, ssl_server_key, tcp_listening_port])
+        return []
 
     def _getOptionByName(self, name):
         """
@@ -138,7 +121,7 @@ class _ConfigurationParser(ConfigParser.ConfigParser, object):
         If no configuration file exists, copy a new default one.
         """
 
-        if not os.path.isfile(self.config_file_location):
+        if len(self.configuration.options) > 0 and not os.path.isfile(self.config_file_location):
             file = open(self.config_file_location, "w")
             file.write(self._generateDefault())
             file.close()
@@ -152,7 +135,8 @@ class _ConfigurationParser(ConfigParser.ConfigParser, object):
 
         @return  (str)    A default configuration file, based on the configuration options.
         """
-        default = '# Gyrid Server configuration file\n[Gyrid Server]\n\n'
+        default = '# Gyrid Server configuration file for %s \n[%s]\n\n' % (self.configuration.name,
+            self.configuration.name)
         for option in self.configuration.options:
             if not option.hidden:
                 default += "\n# ".join(textwrap.wrap("# %s" % option.description, 78))
@@ -176,11 +160,11 @@ class _ConfigurationParser(ConfigParser.ConfigParser, object):
                            option.
         """
         try:
-            return ConfigParser.ConfigParser.get(self, 'Gyrid Server', option)
+            return ConfigParser.ConfigParser.get(self, self.configuration.name, option)
         except:
             return None
 
-class _Option(object):
+class Option(object):
     """
     Class for an option.
     """
